+++ test/test_feature.sh
@@ -70,11 +70,10 @@ if [ $BE -ne 0 ]; then
     fi
 
     new "start backend -s init -f $cfg -y $fyang"
-    # start new backend
-    sudo $clixon_backend -s init -f $cfg -y $fyang -D $DBG
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    start_backend -s init -f $cfg -y $fyang
+
+    new "waiting"
+    sleep $RCWAIT
 fi
 
 new "cli enabled feature"
@@ -183,9 +182,6 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 
 rm -rf $dir