+++ test/test_nacm_module_read.sh
@@ -134,10 +134,7 @@ if [ $BE -ne 0 ]; then
 	err
     fi
     new "start backend -s init -f $cfg"
-    sudo $clixon_backend -s init -f $cfg -D $DBG
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    start_backend -s init -f $cfg
 fi
 
 new "kill old restconf daemon"
@@ -147,6 +144,7 @@ sleep 1
 new "start restconf daemon (-a is enable basic authentication)"
 sudo su -c "$clixon_restconf -f $cfg $RCLOG -D $DBG -- -a" -s /bin/sh www-data &
 
+new "waiting"
 sleep $RCWAIT
 
 new "auth set authentication config"
@@ -287,9 +285,6 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 
 rm -rf $dir