+++ test/test_perf.sh
@@ -4,24 +4,17 @@
 # Magic line must be first in script (see README.md)
 s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi
 
-number=5000
-req=100
-if [ $# = 0 ]; then
-    number=1000
-elif [ $# = 1 ]; then
-    number=$1
-elif [ $# = 2 ]; then
-    number=$1
-    req=$2
-else
-    echo "Usage: $0 [<number> [<requests>]]"
-    exit 1 # Scaling
-fi
+# Number of list/leaf-list entries in file
+: ${perfnr:=1000}
+
+# Number of requests made get/put
+: ${perfreq:=100}
+
 APPNAME=example
 
 cfg=$dir/scaling-conf.xml
 fyang=$dir/scaling.yang
-fconfig=$dir/config
+fconfig=$dir/large.xml
 
 cat <<EOF > $fyang
 module scaling{
@@ -57,10 +50,11 @@ cat <<EOF > $cfg
   <CLICON_RESTCONF_PRETTY>false</CLICON_RESTCONF_PRETTY>
   <CLICON_XMLDB_DIR>/usr/local/var/$APPNAME</CLICON_XMLDB_DIR>
   <CLICON_XMLDB_PLUGIN>/usr/local/lib/xmldb/text.so</CLICON_XMLDB_PLUGIN>
+  <CLICON_XMLDB_PRETTY>false</CLICON_XMLDB_PRETTY>
 </config>
 EOF
 
-new "test params: -f $cfg" -y $fyang
+new "test params: -f $cfg -y $fyang"
 if [ $BE -ne 0 ]; then
     new "kill old backend"
     sudo clixon_backend -zf $cfg -y $fyang
@@ -68,12 +62,8 @@ if [ $BE -ne 0 ]; then
 	err
     fi
 
-    new "start backend  -s init -f $cfg -y $fyang"
-    # start new backend
-    sudo $clixon_backend -s init -f $cfg -y $fyang -D $DBG
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    new "start backend -s init -f $cfg -y $fyang"
+    start_backend -s init -f $cfg -y $fyang
 fi
 
 new "kill old restconf daemon"
@@ -82,69 +72,75 @@ sudo pkill -u www-data -f "/www-data/clixon_restconf"
 new "start restconf daemon"
 sudo su -c "$clixon_restconf -f $cfg -y $fyang $RCLOG -D $DBG" -s /bin/sh www-data &
 
+new "waiting"
 sleep $RCWAIT
 
-new "generate 'large' config with $number list entries"
+new "generate 'large' config with $perfnr list entries"
 echo -n "<rpc><edit-config><target><candidate/></target><config><x xmlns=\"urn:example:clixon\">" > $fconfig
-for (( i=0; i<$number; i++ )); do  
+for (( i=0; i<$perfnr; i++ )); do  
     echo -n "<y><a>$i</a><b>$i</b></y>" >> $fconfig
 done
 echo "</x></config></edit-config></rpc>]]>]]>" >> $fconfig
 
-# Just for manual dbg
-echo "$clixon_netconf -qf $cfg  -y $fyang"
-
+# Now take large config file and write it via netconf to candidate
 new "netconf write large config"
 expecteof_file "/usr/bin/time -f %e $clixon_netconf -qf $cfg -y $fyang" "$fconfig" "^<rpc-reply><ok/></rpc-reply>]]>]]>$"
 
-#echo '<rpc><get-config><source><candidate/></source></get-config></rpc>]]>]]>' | $clixon_netconf -qf $cfg  -y $fyang 
+# Here, there are $perfnr entries in candidate
 
 new "netconf write large config again"
 expecteof_file "/usr/bin/time -f %e $clixon_netconf -qf $cfg -y $fyang" "$fconfig" "^<rpc-reply><ok/></rpc-reply>]]>]]>$"
 
-#echo '<rpc><get-config><source><candidate/></source></get-config></rpc>]]>]]>' | $clixon_netconf -qf $cfg  -y $fyang 
-
+# Remove the file, its used for different purposes further down
 rm $fconfig
 
+# Now commit it from candidate to running 
 new "netconf commit large config"
 expecteof "/usr/bin/time -f %e $clixon_netconf -qf $cfg -y $fyang" 0 "<rpc><commit/></rpc>]]>]]>" "^<rpc-reply><ok/></rpc-reply>]]>]]>$" 
 
+# Now commit it again from candidate (validation takes time when
+# comparing to existing)
 new "netconf commit large config again"
 expecteof "/usr/bin/time -f %e $clixon_netconf -qf $cfg -y $fyang" 0 "<rpc><commit/></rpc>]]>]]>" "^<rpc-reply><ok/></rpc-reply>]]>]]>$" 
 
-new "netconf add small (1 entry) config"
-expecteof "/usr/bin/time -f %e $clixon_netconf -qf $cfg -y $fyang" 0 '<rpc><edit-config><target><candidate/></target><config><x xmlns="urn:example:clixon"><y><a>x</a><b>y</b></y></x></config></edit-config></rpc>]]>]]>' "^<rpc-reply><ok/></rpc-reply>]]>]]>$" 
-
-new "netconf get $req small config"
-time -p for (( i=0; i<$req; i++ )); do
-    rnd=$(( ( RANDOM % $number ) ))
-    echo "<rpc><get-config><source><candidate/></source><filter type=\"xpath\" select=\"/x/y[a=$rnd][b=$rnd]\" /></get-config></rpc>]]>]]>"
-done | $clixon_netconf -qf $cfg  -y $fyang > /dev/null
-
-new "netconf get $req restconf small config"
-time -p for (( i=0; i<$req; i++ )); do
-    rnd=$(( ( RANDOM % $number ) ))
-#XXX    curl -sX PUT -d {"y":{"a":"$rnd","b":"$rnd"}} http://localhost/restconf/data/x/y=$rnd,$rnd 
-done 
-
-new "netconf add $req small config"
-time -p for (( i=0; i<$req; i++ )); do
-    rnd=$(( ( RANDOM % $number ) ))
+# Having a large db, get and put single entries many times
+# Note same entries in the range alreayd there, db has same size
+new "netconf add $perfreq small config"
+time -p for (( i=0; i<$perfreq; i++ )); do
+    rnd=$(( ( RANDOM % $perfnr ) ))
     echo "<rpc><edit-config><target><candidate/></target><config><x xmlns=\"urn:example:clixon\"><y><a>$rnd</a><b>$rnd</b></y></x></config></edit-config></rpc>]]>]]>"
 done | $clixon_netconf -qf $cfg  -y $fyang > /dev/null
 
-new "netconf add $req restconf small config"
-time -p for (( i=0; i<$req; i++ )); do
-    rnd=$(( ( RANDOM % $number ) ))
-    curl -sG http://localhost/restconf/data/x/y=$rnd,$rnd > /dev/null
+new "netconf get $perfreq small config"
+time -p for (( i=0; i<$perfreq; i++ )); do
+    rnd=$(( ( RANDOM % $perfnr ) ))
+    echo "<rpc><get-config><source><candidate/></source><filter type=\"xpath\" select=\"/x/y[a=$rnd][b=$rnd]\" /></get-config></rpc>]]>]]>"
+done | $clixon_netconf -qf $cfg  -y $fyang > /dev/null
+
+new "restconf get $perfreq small config"
+time -p for (( i=0; i<$perfreq; i++ )); do
+    rnd=$(( ( RANDOM % $perfnr ) ))
+    curl -sG http://localhost/restconf/data/scaling:x/y=$rnd,$rnd > /dev/null
+done
+
+new "restconf add $perfreq small config"
+time -p for (( i=0; i<$perfreq; i++ )); do
+    rnd=$(( ( RANDOM % $perfnr ) ))
+    curl -s -X PUT http://localhost/restconf/data/scaling:x/y=$rnd  -d '{"scaling:y":{"a":"'$rnd'","b":"'$rnd'"}}'
 done 
 
+# Instead of many small entries, get one large in netconf and restconf
 new "netconf get large config"
 expecteof "/usr/bin/time -f %e $clixon_netconf -qf $cfg  -y $fyang" 0 "<rpc><get-config><source><candidate/></source></get-config></rpc>]]>]]>" '^<rpc-reply><data><x xmlns="urn:example:clixon"><y><a>0</a><b>0</b></y><y><a>1</a><b>1</b>'
 
+new "restconf get large config"
+expecteof "/usr/bin/time -f %e curl -sG http://localhost/restconf/data" 0 "<rpc><get-config><source><candidate/></source></get-config></rpc>]]>]]>" '^{"data": {"scaling:x": {"y": \[{"a": "0","b": "0"},{ "a": "1","b": "1"},'
+
+# Now do leaf-lists istead of leafs
+
 new "generate large leaf-list config"
 echo -n "<rpc><edit-config><target><candidate/></target><default-operation>replace</default-operation><config><x xmlns=\"urn:example:clixon\">" > $fconfig
-for (( i=0; i<$number; i++ )); do  
+for (( i=0; i<$perfnr; i++ )); do  
     echo -n "<c>$i</c>" >> $fconfig
 done
 echo "</x></config></edit-config></rpc>]]>]]>" >> $fconfig
@@ -157,9 +153,9 @@ rm $fconfig
 new "netconf commit large leaf-list config"
 expecteof "/usr/bin/time -f %e $clixon_netconf -qf $cfg -y $fyang" 0 "<rpc><commit/></rpc>]]>]]>" "^<rpc-reply><ok/></rpc-reply>]]>]]>$" 
 
-new "netconf add $req small leaf-list config"
-time -p for (( i=0; i<$req; i++ )); do
-    rnd=$(( ( RANDOM % $number ) ))
+new "netconf add $perfreq small leaf-list config"
+time -p for (( i=0; i<$perfreq; i++ )); do
+    rnd=$(( ( RANDOM % $perfnr ) ))
     echo "<rpc><edit-config><target><candidate/></target><config><x xmlns=\"urn:example:clixon\"><c>$rnd</c></x></config></edit-config></rpc>]]>]]>"
 done | $clixon_netconf -qf $cfg  -y $fyang > /dev/null
 
@@ -186,9 +182,6 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 
 rm -rf $dir