+++ test/test_startup.sh
@@ -36,7 +36,7 @@ cat <<EOF > $cfg
 
 EOF
 
-run(){
+testrun(){
     mode=$1
     expect=$2
 
@@ -83,11 +83,11 @@ EOF
 	err
     fi
 
-    new "start backend  -f $cfg -s $mode -c $dir/config"
-    sudo $clixon_backend -f $cfg -s $mode -c $dir/config
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    new "start backend -f $cfg -s $mode -c $dir/config"
+    start_backend -s $mode -f $cfg -c $dir/config
+
+    new "waiting"
+    sleep $RCWAIT
 
     new "Check $mode"
     expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><get-config><source><running/></source></get-config></rpc>]]>]]>' "^<rpc-reply>$expect</rpc-reply>]]>]]>$"
@@ -99,15 +99,15 @@ EOF
 	err "backend already dead"
     fi
     # kill backend
-    sudo clixon_backend -z -f $cfg
-    if [ $? -ne 0 ]; then
-	err "kill backend"
-    fi
-}
+    stop_backend -f $cfg
+} # testrun
 
-run init    '<data/>'
-run none    '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>run</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
-run running '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>extra</name><type>ex:eth</type><enabled>true</enabled></interface><interface><name>lo</name><type>ex:loopback</type><enabled>true</enabled></interface><interface><name>run</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
-run startup '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>extra</name><type>ex:eth</type><enabled>true</enabled></interface><interface><name>lo</name><type>ex:loopback</type><enabled>true</enabled></interface><interface><name>startup</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
+testrun init    '<data/>'
+
+testrun none    '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>run</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
+
+testrun running '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>extra</name><type>ex:eth</type><enabled>true</enabled></interface><interface><name>lo</name><type>ex:loopback</type><enabled>true</enabled></interface><interface><name>run</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
+
+testrun startup '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>extra</name><type>ex:eth</type><enabled>true</enabled></interface><interface><name>lo</name><type>ex:loopback</type><enabled>true</enabled></interface><interface><name>startup</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
 
 rm -rf $dir