+++ test/test_stream.sh
@@ -111,10 +111,7 @@ if [ $BE -ne 0 ]; then
 	err
     fi
     new "start backend -s init -f $cfg -y $fyang"
-    sudo $clixon_backend -s init -f $cfg -y $fyang -D $DBG
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    start_backend -s init -f $cfg -y $fyang
 fi
 
 new "kill old restconf daemon"
@@ -123,6 +120,7 @@ sudo pkill -u www-data -f "/www-data/clixon_restconf"
 new "start restconf daemon"
 sudo su -c "$clixon_restconf -f $cfg -y $fyang $RCLOG -D $DBG" -s /bin/sh www-data &
 
+new "waiting"
 sleep $RCWAIT
 
 #
@@ -289,9 +287,6 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 
-#rm -rf $dir
+rm -rf $dir