+++ test/test_union.sh
@@ -86,10 +86,10 @@ if [ $BE -ne 0 ]; then
 	err
     fi
     new "start backend -s init -f $cfg -y $fyang"
-    sudo $clixon_backend -s init -f $cfg -y $fyang
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    start_backend -s init -f $cfg -y $fyang
+
+    new "waiting"
+    sleep $RCWAIT
 fi
 
 new "cli set transitive string"
@@ -112,9 +112,6 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 
 rm -rf $dir