+++ test/test_yang.sh
@@ -151,10 +151,10 @@ if [ $BE -ne 0 ]; then
 	err
     fi
     new "start backend -s init -f $cfg -y $fyang"
-    sudo $clixon_backend -s init -f $cfg -y $fyang -D $DBG
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    start_backend -s init -f $cfg -y $fyang
+
+    new "waiting"
+    sleep $RCWAIT
 fi
 
 new "cli defined extension"
@@ -292,10 +292,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 rm -rf $dir