+++ test/test_yang_load.sh
@@ -79,10 +79,10 @@ if [ $BE -ne 0 ]; then
 	err
     fi
     new "start backend  -s init -f $cfg"
-    sudo $clixon_backend -s init -f $cfg -D $DBG
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    start_backend -s init -f $cfg
+
+    new "waiting"
+    sleep $RCWAIT
 fi
 
 new "1. Set newex"
@@ -105,10 +105,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 #--------------------------------------
@@ -133,10 +130,10 @@ EOF
 
 new "start backend  -s init -f $cfg"
 # start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set oldex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><oldex xmlns="urn:example:clixon">str</oldex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -154,10 +151,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 #--------------------------------------
@@ -177,11 +171,10 @@ cat <<EOF > $cfg
 EOF
 
 new "start backend  -s init -f $cfg"
-# start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set newex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><newex xmlns="urn:example:clixon">str</newex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -198,11 +191,7 @@ pid=`pgrep -u root -f clixon_backend`
 if [ -z "$pid" ]; then
     err "backend already dead"
 fi
-# kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 #--------------------------------------
@@ -223,11 +212,10 @@ cat <<EOF > $cfg
 EOF
 
 new "start backend  -s init -f $cfg"
-# start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set oldex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><oldex xmlns="urn:example:clixon">str</oldex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -245,10 +233,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 #--------------------------------------
@@ -268,11 +253,10 @@ cat <<EOF > $cfg
 EOF
 
 new "start backend  -s init -f $cfg"
-# start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set newex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><newex xmlns="urn:example:clixon">str</newex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -290,10 +274,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 #--------------------------------------
@@ -314,11 +295,10 @@ cat <<EOF > $cfg
 EOF
 
 new "start backend  -s init -f $cfg"
-# start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set oldex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><oldex xmlns="urn:example:clixon">str</oldex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -336,10 +316,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 
@@ -362,11 +339,10 @@ cat <<EOF > $cfg
 EOF
 
 new "start backend  -s init -f $cfg"
-# start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set oldex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><oldex xmlns="urn:example:clixon">str</oldex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -384,10 +360,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 #--------------------------------------
@@ -409,11 +382,10 @@ cat <<EOF > $cfg
 EOF
 
 new "start backend  -s init -f $cfg"
-# start new backend
-sudo $clixon_backend -s init -f $cfg
-if [ $? -ne 0 ]; then
-    err
-fi
+start_backend -s init -f $cfg
+
+new "waiting"
+sleep $RCWAIT
 
 new "Set oldex"
 expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><edit-config><target><candidate/></target><config><oldex xmlns="urn:example:clixon">str</oldex></config></edit-config></rpc>]]>]]>' '^<rpc-reply><ok/></rpc-reply>]]>]]>$'
@@ -431,10 +403,7 @@ if [ -z "$pid" ]; then
     err "backend already dead"
 fi
 # kill backend
-sudo clixon_backend -z -f $cfg
-if [ $? -ne 0 ]; then
-    err "kill backend"
-fi
+stop_backend -f $cfg
 sudo pkill -u root -f clixon_backend
 
 rm -rf $dir